// src/features/cart/cart-type.ts
export type cartDataType = {
  brand: string;
  description: string;
  discountPercentage: number;
  id: number;
  price: number;
  quantity: number;
  rating: number;
  thumbnail: string;
  title: string;
};

export type cartRawDataType = {
  brand: string;
  description: string;
  discountPercentage: number;
  id: number;
  price: number;
  quantity: number;
  rating: number;
  thumbnail: string;
  title: string;
};